requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
psycopg2==2.8.6
pycparser==2.20
python-environ==0.4.54
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.4
whitenoise==5.2.0
